Group common HTJ2K functions
